requirements.txt
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
cryptography>=44.0.2
h11==0.14.0
idna==3.10
keyboard==0.13.5
MouseInfo==0.1.3
outcome==1.3.0.post0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytweening==1.2.0
selenium==4.31.0
setuptools==78.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.13.1
urllib3==2.3.0
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0

# Adicionados para o bot do Discord e ambiente do Render
discord.py>=2.3.2
python-dotenv>=1.0.0